apps/processor/requirements.txt
fastapi
uvicorn
redis
python-dotenv
langchain
langchain-openai
langchain-community
langchain-chroma
langchain-text-splitters
chromadb
pypdf
python-docx
supabase
prisma
httpx
python-multipart
numpy>=2.1.0
pydantic
pydantic-settings
pydantic-extra-types
google-generativeai
langchain-google-genai
